refactor(mdorim): tranlations minor refactor
