+++ packages/mdorim/src/translations/index.ts
@@ -1,8 +1,2 @@
-import pt_br from "./pt-br.json" with { type: "json" };
-
-export const translations = {
-    "pt-br": pt_br,
-    "en-us": {},
-};
-
 export * from "./I18n";
+export * from "./translations";
