+++ packages/mdorim/src/translations/translations.ts
@@ -1,0 +1,6 @@
+import pt_br from "./pt-br.json" with { type: "json" };
+
+export const translations = {
+    "pt-br": pt_br,
+    "en-us": {},
+};
